moved ErrorFallback component to pages folder
